Fix sync script
Now that redwood uses vite for the web side in dev instead of webpack,
we needed to change the way the `yarn sync` script works to be
compatible with vite.

This PR adds a vite plugin that tells vite to also look in
`node_modules/@redwoodjs-stripe/...` when watching. It also builds both
cjs and esm outputs for api and web, except we still only use cjs as
the module type for both for now. The esm outputs are only used when
this vite plugin is enabled.

Once this is merged and released the contribution guide will also be
updated to include using this plugin.